[#2] Refactor: 사용하던 파일이름들을 kebab-case로 통일성 있게 변경
- 파일이름들은 모두 Kebab-case로 변경하는 이유는 .tsx파일에서 camelCase와 PascalCase가 혼용되서 사용되고 있기 때문입니다.
- 모두 소문자로 쓰고 단어간 공백은 하이픈을 써서 표현하는게 더 편하기 때문에 kebab-case로 폴더 이름을 바꾸기로 하였습니다
